[www] Add user paywall verification route. (#232)
This new route verifies whether the given transaction is found on the
blockchain, is sent to the correct address, has the minimum number of
confirmations (2), comes after a fixed timestamp, and is of at least
the minimum amount of DCR (0.1). If all those are true, it sets the
transaction on the user in the database, marking the user as paid.

+++ politeiawww/cmd/politeiawww_refclient/client.go
@@ -187,7 +187,7 @@ func idFromString(s string) (*identity.FullIdentity, error) {
 	return id, nil
 }
 
-func (c *ctx) newUser(email string, password string) (string, *identity.FullIdentity, string, float64, error) {
+func (c *ctx) newUser(email string, password string) (string, *identity.FullIdentity, string, uint64, error) {
 	id, err := idFromString(email)
 	if err != nil {
 		return "", nil, "", 0, err
